Move salt-minion service handling into MSI
During testing there were several instances of 'net start' failing to
start the salt-minion service with no useful error produces visually or
in the log. This change moves the service handling into the Wix
Component (see <ServiceControl /> in service.wxs).

This had a cascade effect on when the configuration manipulation
actions run, and how the user configuration selections are passed to
those actions.

The 'start the service?' checkbox was moved from the exit dialog to the
settings customization dialog.

Also fix build configuration platform setting for Debug|amd64.